Stop calculating sun events twice.
After feedback in the issue
https://github.com/shred/commons-suncalc/issues/30 I have decided to
stop calculating all sun events twice (to work around differing results
depending on the input date) and instead truncate the results to minutes
instead of seconds.

This should give some performance improvement, and will also better
reflect the actual precision of the calculations.

I've added some additional tests to verify that the same minute is
the result no matter when during the day the input date is. Some
additional tests could be added in the future.