Change the behavior when the sun doesn't set.
Instead of simply dropping down or up to 65.7 or -65.7, go towards the
equator in increments starting with whole degrees and smaller the closer
to 65.7 or -65.7 it gets, until an observable sunset is found. This gets
much closer to the actual location and gives a better approximation on
what time it is.

This could be further refined by trying to go up and down until the
closest possible sunset is found, but calculations in the polar regions
are kind of an edge case to start with, and this calculation is heavy
enough as it is. Still, there is room for improvement.

Also, return :lat, :lon, and :date from `next-sunset` to make it
possible to reason about the results.